Update Program.cs
In the current code, the Length of the sourceBlob is not set - and this causes the code to crash. You must first call FetchAttributes() so I have added this in. 
With this and my previous pull request, this code should be working 'out of the box' again...
Thanks for providing it